requirements.txt
pandasai>=3.0.0b2
openai==1.79.0
python-dotenv==1.1.0
pandasai-openai
